Installing Electrum dependency
Electrum 3.2.2 depends on a lib called Typing, which is available in the 
Python 3.5 stdlib. Ideally we should upgrade Python to 3.5 or later, but 
that could break other things and should be tested properly.

For now, let's just manually install the dependency until Python is 
upgraded.

More info: https://pypi.org/project/typing/